Create Storage class to retrive and store tasks
